fix: Corrected SpringBootTests for GraphQLController auto-configuratino
