+++ graphql-jpa-query-web/src/test/java/com/introproventures/graphql/jpa/query/test/web/autoconfigure/GraphQLControllerAutoConfigurationTest.java
@@ -1,7 +1,9 @@
-package com.introproventures.graphql.jpa.query.web.autoconfigure;
+package com.introproventures.graphql.jpa.query.test.web.autoconfigure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.introproventures.graphql.jpa.query.schema.GraphQLExecutor;
+import com.introproventures.graphql.jpa.query.web.GraphQLController;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,18 +13,14 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import com.introproventures.graphql.jpa.query.schema.GraphQLExecutor;
-import com.introproventures.graphql.jpa.query.web.GraphQLController;
-
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment=WebEnvironment.RANDOM_PORT,
-                properties = "spring.graphql.jpa.query.enabled=false")
-public class GraphQLControllerAutoConfigurationPropertyDisabledTest {
+@SpringBootTest(webEnvironment=WebEnvironment.RANDOM_PORT)
+public class GraphQLControllerAutoConfigurationTest {
     
     @MockBean
     private GraphQLExecutor graphQLExecutor;
     
-    @Autowired(required=false)
+    @Autowired
     private GraphQLController graphQLController;
     
     @SpringBootApplication
@@ -32,7 +30,7 @@ public class GraphQLControllerAutoConfigurationPropertyDisabledTest {
     
     @Test
     public void contextLoads() {
-        assertThat(graphQLController).isNull();
+        assertThat(graphQLController).isNotNull();
     }
 
 }
